move authentication to separate concern; scope for session params
