+++ spec/requests/api/v1/sessions_request_spec.rb
@@ -7,17 +7,19 @@ RSpec.describe "Api::V1::Sessions", type: :request do
     let(:credentials) { { email: "email@sessions.spec", password: "123456" } }
     let(:headers) { { "Content-Type" => "application/json" } }
     let(:do_request) do
-      post "/api/v1/sessions", params: credentials.to_json, headers: headers
+      post "/api/v1/sessions", params: { session: credentials }.to_json, headers: headers
     end
 
-    before { create(:user, credentials) }
+    context "with valid credentials" do
+      before { create(:user, credentials) }
 
-    it "responds with token" do
-      expect(response_body).to match("token" => match(/^eyJ.+\..+\..+$/))
+      it "responds with token" do
+        expect(response_body).to match("token" => match(/^eyJ.+\..+\..+$/))
+      end
     end
 
     context "with invalid credentials" do
-      let(:credentials) { {} }
+      let(:credentials) { { email: "fake@email.spec", password: "fake_pw" } }
 
       it "responds with error" do
         expect(response_body).to match("error" => include("Invalid credentials"))
